% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genSWATdates.R
\name{genSWATdates}
\alias{genSWATdates}
\title{generates a vector of dates in the swat model IO format}
\usage{
genSWATdates(startDate, endDate)
}
\arguments{
\item{startDate}{}

\item{endDate}{}
}
\value{
returns a generated time series of swat IO format dates
}
\description{
generates a vector of dates in the swat model IO format
}

