% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getMet.R
\name{getMet}
\alias{getMet}
\title{Gets met data from a specified source and creates model input files in the specified format}
\usage{
getMet(locations, dataSource = "cfsr", outFormat = "swat",
  outDir = getwd())
}
\arguments{
\item{locations}{}

\item{dataSource}{}

\item{outFormat}{}

\item{outDir}{}
}
\value{
returns specified  met data in specified format
}
\description{
Gets met data from a specified source and creates model input files in the specified format
}

