% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSWATcfsr.R
\name{getSWATcfsr}
\alias{getSWATcfsr}
\title{Gets met data from the redimentioned CFSR data set and outputs SWAT IO format weather input files}
\usage{
getSWATcfsr(centroids, outDir = getwd())
}
\arguments{
\item{centroids}{}

\item{outDir}{}
}
\value{
returns cfsr met data in swat IO format
}
\description{
Gets met data from the redimentioned CFSR data set and outputs SWAT IO format weather input files
}

