% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSWATdates.R
\name{getSWATdates}
\alias{getSWATdates}
\title{Converts a date time series to SWAT IO format dates}
\usage{
getSWATdates(dates, dateformat = "yyyy-mm-dd")
}
\arguments{
\item{dates}{}

\item{dateformat}{}
}
\value{
returns a time series of swat IO format dates from the input
}
\description{
Converts a date time series to SWAT IO format dates
}

